+++ packages/dso-toolkit/src/components/map-message/map-message.models.ts
@@ -1,0 +1,8 @@
+import { Button } from "../button/button.models.js";
+
+export interface MapMessage {
+  variant: "success" | "error" | "instruction";
+  message: string;
+  showButtons?: boolean;
+  buttons?: Button[];
+}