+++ storybook/src/templates.ts
@@ -70,6 +70,7 @@ import {
   Logo,
   MapControls,
   MapLayer,
+  MapMessage,
   MarkBar,
   Modal,
   Navbar,
@@ -107,6 +108,7 @@ import {
 import { TemplateResult } from "lit-html";
 
 // Start: import templates
+
 import { coreAccordion } from "./components/accordion/accordion.core-template";
 import { coreActionList } from "./components/action-list/action-list.core-template";
 import { coreAdvancedSelect } from "./components/advanced-select/advanced-select.core-template";
@@ -182,6 +184,7 @@ import { coreListButton } from "./components/list-button/list-button.core-templa
 import { coreLogo } from "./components/logo/logo.core-template";
 import { coreMapControls } from "./components/map-controls/map-controls.core-template";
 import { coreMapLayer } from "./components/map-layer/map-layer.core-template";
+import { coreMapMessage } from "./components/map-message/map-message.core-template";
 import { coreMarkBar } from "./components/mark-bar/mark-bar.core-template";
 import { coreModal } from "./components/modal/modal.core-template";
 import { cssModal } from "./components/modal/modal.css-template";
@@ -214,7 +217,6 @@ import { coreToggletip } from "./components/toggletip/toggletip.core-template";
 import { coreTooltip } from "./components/tooltip/tooltip.core-template";
 import { coreTreeView } from "./components/tree-view/tree-view.core-template";
 import { coreViewerGrid } from "./components/viewer-grid/viewer-grid.core-template";
-
 // End: import templates
 
 export interface Components {
@@ -286,6 +288,7 @@ export interface Components {
   logo: Logo;
   mapControls: MapControls;
   mapLayer: MapLayer<TemplateResult>;
+  mapMessage: MapMessage;
   markBar: MarkBar;
   modal: Modal<TemplateResult>;
   navbar: Navbar<TemplateResult>;
@@ -361,6 +364,7 @@ templateContainer.add(coreButtonGroup);
 templateContainer.add(coreCard);
 templateContainer.add(coreCardContainer);
 templateContainer.add(coreContactInformation);
+templateContainer.add(coreCursorTooltip);
 templateContainer.add(coreDatePicker);
 templateContainer.add(coreDocumentCard);
 templateContainer.add(coreDocumentComponent);
@@ -381,6 +385,7 @@ templateContainer.add(coreListButton);
 templateContainer.add(coreLogo);
 templateContainer.add(coreMapControls);
 templateContainer.add(coreMapLayer);
+templateContainer.add(coreMapMessage);
 templateContainer.add(coreMarkBar);
 templateContainer.add(coreModal);
 templateContainer.add(coreOnboardingTip);
@@ -454,5 +459,4 @@ templateContainer.add(cssSearchBar);
 templateContainer.add(cssShoppingCart);
 templateContainer.add(cssTile);
 templateContainer.add(cssTileGrid);
-templateContainer.add(coreCursorTooltip);
 // End: add templates to container